refactor: Extract delayed debit specs
